pyproject.toml
[tool.poetry]
name = "discord-bot-v3"
version = "0.1.0"
description = ""
authors = ["Zach Fuller <zach.fuller1222@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_bot_v3"}]

[tool.poetry.dependencies]
python = "^3.11"
pynacl = "^1.5.0"
brotlipy = "^0.7.0"
orjson = "^3.8.9"
aiodns = "^3.0.0"
python-dotenv = "^1.0.0"
py-cord = {extras = ["voice"], version = "^2.4.1"}
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
